[ADD] Login상태일 때만 header가 보이도록 수정, Outlet 사용
