+++ src/pages/home/Home.jsx
@@ -1,11 +1,9 @@
-import Layout from "@/components/commons/layout/Layout";
 import HomeBody from "../../components/units/homeComponent/homeBody/HomeBody";
 import HomeHeader from "../../components/units/homeComponent/homeHeader/HomeHeader";
 
 function Home() {
   return (
     <>
-      <Layout />
       <HomeHeader />
       <HomeBody />
     </>
